Force UTF-8 encoding on the response
When I fetch the YAML using JRuby 1.7.1 and 1.7.2
I get the response body in ASCII-8BIT and the file content
is broken. This differs from the CRuby behavior where I will
get the file in correct encoding.

This is an issue with RestClient. The behavior differs only for
HTTP responses that contain file attachments. In these cases
running on JRuby the RestClient will return ASCII-8BIT. I couldn't
isolate it to Net::HTTP, so I'm unsure where the actual problem is.

I'm forcing the UTF-8 as I think we *always* expect UTF-8 as response
for the YAML files.